requirements.txt
nbtlib==2.0.4
pillow~=11.3.0